Move string_index into separate file
