+++ libvast/CMakeLists.txt
@@ -120,6 +120,7 @@ set(libvast_sources
     src/icmp.cpp
     src/ids.cpp
     src/index/enumeration_index.cpp
+    src/index/string_index.cpp
     src/io/read.cpp
     src/io/save.cpp
     src/io/write.cpp
@@ -340,6 +341,7 @@ set(tests
     test/ids.cpp
     test/index/arithmetic_index.cpp
     test/index/enumeration_index.cpp
+    test/index/string_index.cpp
     test/iterator.cpp
     test/json.cpp
     test/meta_index.cpp
